+++ example/lib/main.dart
@@ -132,7 +132,7 @@ class _MainAppState extends State<MainApp> {
           title: const Text('Flutter Data Grid'),
           actions: [_buildToolbar()],
         ),
-        body: DataGrid<ProductRow>(controller: controller, cacheExtent: 960),
+        body: DataGrid<ProductRow>(controller: controller),
       ),
     );
   }